Test API iterativelly
