+++ manager/ui_framework/tests/tests_api.py
@@ -1,27 +1,104 @@
 """Test the UI Framework API."""
-import json
 from django.contrib.auth.models import User, Permission
-from django.test import TestCase
 from django.urls import reverse
-from django.utils import timezone
-from freezegun import freeze_time
-from rest_framework import serializers, status
-from rest_framework.test import APIClient
+from rest_framework import status
 from api.models import Token
-from ui_framework.models import Workspace, View, WorkspaceView
+from ui_framework.tests.utils import get_dict, BaseTestCase
 
 
-class WorkspaceCrudTestCase(TestCase):
-    """Test the workspace CRUD API."""
+class UnauthenticatedCrudTestCase(BaseTestCase):
+    """Test that unauthenticated users cannot use the CRUD API."""
 
     def setUp(self):
-        """Testcase setup.
-
-        We start with 3 workspaces and 4 views and we add view_i and view_i+1 to workspace_i
-        """
+        """Set testcase. Inherits from utils.BaseTestCase."""
         # Arrange
-        # Client setup
-        self.client = APIClient()
+        super().setUp()
+
+    def test_unauthenticated_list_objects(self):
+        """Test that unauthenticated users cannot retrieve the list of objects through the API."""
+        for case in self.cases:
+            # Act
+            url = reverse('{}-list'.format(case['key']))
+            response = self.client.get(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_401_UNAUTHORIZED,
+                'Get list of {} did not return status 401'.format(case['class'])
+            )
+
+    def test_unauthenticated_create_objects(self):
+        """Test that unauthenticated users cannot create objects through the API."""
+        for case in self.cases:
+            # Act
+            url = reverse('{}-list'.format(case['key']))
+            response = self.client.post(url, case['new_data'])
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_401_UNAUTHORIZED,
+                'Posting a new {} did not return status 401'.format(case['class'])
+            )
+            self.assertEqual(
+                case['class'].objects.count(), case['old_count'],
+                'The number of {} should not have changed'.format(case['class'])
+            )
+
+    def test_unauthenticated_retrieve_objects(self):
+        """Test that unauthenticated users cannot retrieve objects through the API."""
+        for case in self.cases:
+            # Act
+            obj = case['class'].objects.first()
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.get(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_401_UNAUTHORIZED,
+                'Getting a {} did not return status 401'.format(case['class'])
+            )
+
+    def test_unauthenticated_update_objects(self):
+        """Test that unauthenticated users cannot update objects through the API."""
+        for case in self.cases:
+            # Act
+            obj = case['class'].objects.first()
+            old_data = get_dict(obj)
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.put(url, case['new_data'])
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_401_UNAUTHORIZED,
+                'Updating a {} did not return status 401'.format(case['class'])
+            )
+            new_data = get_dict(case['class'].objects.get(pk=obj.pk))
+            self.assertEqual(
+                new_data, old_data,
+                'The object {} should not have been updated'.format(case['class'])
+            )
+
+    def test_unauthenticated_delete_objects(self):
+        """Test that unauthenticated users cannot dalete objects through the API."""
+        for case in self.cases:
+            # Act
+            obj = case['class'].objects.first()
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.delete(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_401_UNAUTHORIZED,
+                'Deleting a {} did not return status 401'.format(case['class'])
+            )
+            self.assertEqual(
+                case['class'].objects.count(), case['old_count'],
+                'The number of {} should not have changed'.format(case['class'])
+            )
+
+
+class UnauthorizedCrudTestCase(BaseTestCase):
+    """Test that unauthorized users cannot use the CRUD API."""
+
+    def setUp(self):
+        """Set testcase. Inherits from utils.BaseTestCase."""
+        # Arrange
+        super().setUp()
         self.login_url = reverse('login')
         self.username = 'test'
         self.password = 'password'
@@ -32,177 +109,106 @@ class WorkspaceCrudTestCase(TestCase):
             first_name='First',
             last_name='Last',
         )
-        self.setup_ts = timezone.now()
-        self.setup_ts_str = serializers.DateTimeField().to_representation(self.setup_ts)
-
-        with freeze_time(self.setup_ts):
-            # Data to be used to create views and workspaces
-            self.views_data = [
-                {
-                    'name': 'My View 1',
-                    'data': json.dumps({"data_name": "My View 1"}),
-                },
-                {
-                    'name': 'My View 2',
-                    'data': json.dumps({"data_name": "My View 2"}),
-                },
-                {
-                    'name': 'My View 3',
-                    'data': json.dumps({"data_name": "My View 3"}),
-                },
-                {
-                    'name': 'My View 4',
-                    'data': json.dumps({"data_name": "My View 4"}),
-                }
-            ]
-            self.workspaces_data = [
-                {'name': 'My Workspace 1'},
-                {'name': 'My Workspace 2'},
-                {'name': 'My Workspace 3'},
-            ]
-            self.views = []
-            self.workspaces = []
-            # Create views and store them in self.views
-            for i in range(0, len(self.views_data)):
-                self.views.append(View.objects.create(**self.views_data[i]))
-
-            # Create workspaces, add them the views and store them in self.workspaces
-            for i in range(0, len(self.workspaces_data)):
-                aux = Workspace.objects.create(**self.workspaces_data[i])
-                self.workspaces_data[i]['id'] = aux.id
-                self.workspaces.append(aux)
-                self.workspaces[i].views.add(self.views[i], through_defaults={'view_name': 'v{}'.format(i)})
-                self.workspaces[i].views.add(self.views[i + 1], through_defaults={'view_name': 'v{}'.format(i)})
-        self.old_count = Workspace.objects.count()
-
-    def client_login(self):
-        """Perform a login for the APIClient."""
         data = {'username': self.username, 'password': self.password}
         self.client.post(self.login_url, data, format='json')
         self.token = Token.objects.get(user__username=self.username)
         self.client.credentials(HTTP_AUTHORIZATION='Token ' + self.token.key)
 
-    def test_list_workspaces(self):
-        """Test that the list of workspaces can be retrieved through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='view_workspace'))
+    def test_unauthorized_list_objects(self):
+        """Test that unauthorized users can still retrieve the list of objects through the API."""
+        for case in self.cases:
+            # Act
+            url = reverse('{}-list'.format(case['key']))
+            response = self.client.get(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_200_OK,
+                'Retrieving list of {} did not return status 200'.format(case['class'])
+            )
+            retrieved_data = [dict(data) for data in response.data]
+            self.assertEqual(
+                retrieved_data, case['current_data'],
+                'Retrieved list of {} is not as expected'.format(case['class'])
+            )
 
-        # Act
-        response = self.client.get(reverse('workspace-list'))
+    def test_unauthorized_create_objects(self):
+        """Test that unauthorized users cannot create objects through the API."""
+        for case in self.cases:
+            # Act
+            url = reverse('{}-list'.format(case['key']))
+            response = self.client.post(url, case['new_data'])
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_403_FORBIDDEN,
+                'Posting a new {} did not return status 403'.format(case['class'])
+            )
+            self.assertEqual(
+                case['class'].objects.count(), case['old_count'],
+                'The number of {} should not have changed'.format(case['class'])
+            )
 
-        # Assert
-        expected_data = []
-        for i in range(0, len(self.workspaces_data)):
-            expected_data.append({
-                'id': self.workspaces_data[i]['id'],
-                'name': self.workspaces_data[i]['name'],
-                'creation_timestamp': self.setup_ts_str,
-                'update_timestamp': self.setup_ts_str,
-                'views': [v.pk for v in self.views[i: i + 2]],
-            })
+    def test_unauthorized_retrieve_objects(self):
+        """Test that unauthorized users can still retrieve objects through the API."""
+        for case in self.cases:
+            # Act
+            obj = case['class'].objects.get(id=case['selected_id'])
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.get(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_200_OK,
+                'Getting a {} did not return status 200'.format(case['class'])
+            )
+            retrieved_data = dict(response.data)
+            self.assertEqual(
+                retrieved_data, case['current_data'][0],
+                'Retrieved list of {} is not as expected'.format(case['class'])
+            )
 
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = [dict(w) for w in response.data]
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
+    def test_unauthorized_update_objects(self):
+        """Test that unauthorized users cannot update objects through the API."""
+        for case in self.cases:
+            # Act
+            obj = case['class'].objects.get(id=case['selected_id'])
+            old_data = get_dict(obj)
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.put(url, case['new_data'])
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_403_FORBIDDEN,
+                'Updating a {} did not return status 403'.format(case['class'])
+            )
+            new_data = get_dict(case['class'].objects.get(pk=obj.pk))
+            self.assertEqual(
+                new_data, old_data,
+                'The object {} should not have been updated'.format(case['class'])
+            )
 
-    def test_create_workspaces(self):
-        """Test that a workspace can be created through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='add_workspace'))
-        given_data = {
-            'name': 'My New Workspace',
-            'views': [],
-        }
-
-        # Act
-        response = self.client.post(reverse('workspace-list'), given_data)
-
-        # Assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED, 'The request failed')
-        self.new_count = Workspace.objects.count()
-        self.assertEqual(self.old_count + 1, self.new_count, 'A new object should have been created in the DB')
-        new_workspace = Workspace.objects.get(name=given_data['name'])
-        new_workspace_views = [v.pk for v in new_workspace.views.all()]
-        self.assertEqual(new_workspace_views, given_data['views'], 'Retrieved views are not as expected')
-
-    def test_retrieve_workspaces(self):
-        """Test that a workspace can be retrieved through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='view_workspace'))
-        workspace_data = self.workspaces_data[0]
-        # Act
-        response = self.client.get(reverse('workspace-detail', kwargs={'pk': workspace_data['id']}))
-
-        # Assert
-        expected_data = {
-            'id': workspace_data['id'],
-            'name': workspace_data['name'],
-            'creation_timestamp': self.setup_ts_str,
-            'update_timestamp': self.setup_ts_str,
-            'views': [v.pk for v in self.views[0:2]],
-        }
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = dict(response.data)
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
-
-    def test_update_workspaces(self):
-        """Test that a workspace can be updated through the API."""
-        # Arrange
-        self.client_login()
-        workspace = self.workspaces[0]
-        self.user.user_permissions.add(Permission.objects.get(codename='change_workspace'))
-        given_data = {
-            'name': 'My New Workspace',
-        }
-        # Act
-        self.update_ts = timezone.now()
-        self.update_ts_str = serializers.DateTimeField().to_representation(self.update_ts)
-        with freeze_time(self.update_ts):
-            response = self.client.put(reverse('workspace-detail', kwargs={'pk': workspace.pk}), given_data)
-
-        # Assert
-        expected_data = {
-            'id': workspace.id,
-            'name': given_data['name'],
-            'creation_timestamp': self.setup_ts_str,
-            'update_timestamp': self.update_ts_str,
-            'views': [v.pk for v in self.views[0:2]],
-        }
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = dict(response.data)
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
-
-    def test_delete_workspaces(self):
-        """Test that a workspace can be deleted through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='delete_workspace'))
-        workspace_pk = self.workspaces[0].pk
-        # Act
-        response = self.client.delete(reverse('workspace-detail', kwargs={'pk': workspace_pk}))
-
-        # Assert
-        self.new_count = Workspace.objects.count()
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT, 'The request failed')
-        self.assertEqual(self.old_count - 1, self.new_count, 'The number of objects in the DB have decreased by 1')
-        with self.assertRaises(Exception):
-            Workspace.objects.get(pk=workspace_pk)
+    def test_unauthorized_delete_objects(self):
+        """Test that unauthorized users cannot dalete objects through the API."""
+        for case in self.cases:
+            # Act
+            obj = case['class'].objects.first()
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.delete(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_403_FORBIDDEN,
+                'Deleting a {} did not return status 403'.format(case['class'])
+            )
+            self.assertEqual(
+                case['class'].objects.count(), case['old_count'],
+                'The number of {} should not have changed'.format(case['class'])
+            )
 
 
-class ViewCrudTestCase(TestCase):
-    """Test the view CRUD API."""
+class AuthorizedCrudTestCase(BaseTestCase):
+    """Test that authorized users can use the CRUD API."""
 
     def setUp(self):
-        """Testcase setup.
-
-        We start with 3 workspaces and 4 views and we add view_i and view_i+1 to workspace_i
-        """
+        """Set testcase. Inherits from utils.BaseTestCase."""
         # Arrange
-        # Client setup
-        self.client = APIClient()
+        super().setUp()
         self.login_url = reverse('login')
         self.username = 'test'
         self.password = 'password'
@@ -213,361 +219,105 @@ class ViewCrudTestCase(TestCase):
             first_name='First',
             last_name='Last',
         )
-        self.setup_ts = timezone.now()
-        self.setup_ts_str = serializers.DateTimeField().to_representation(self.setup_ts)
-
-        with freeze_time(self.setup_ts):
-            # Data to be used to create views and workspaces
-            self.views_data = [
-                {
-                    'name': 'My View 1',
-                    'data': json.dumps({"data_name": "My View 1"}),
-                },
-                {
-                    'name': 'My View 2',
-                    'data': json.dumps({"data_name": "My View 2"}),
-                },
-                {
-                    'name': 'My View 3',
-                    'data': json.dumps({"data_name": "My View 3"}),
-                },
-                {
-                    'name': 'My View 4',
-                    'data': json.dumps({"data_name": "My View 4"}),
-                }
-            ]
-            self.workspaces_data = [
-                {'name': 'My Workspace 1'},
-                {'name': 'My Workspace 2'},
-                {'name': 'My Workspace 3'},
-            ]
-            self.views = []
-            self.workspaces = []
-            # Create views and store them in self.views
-            for i in range(0, len(self.views_data)):
-                aux = View.objects.create(**self.views_data[i])
-                self.views_data[i]['id'] = aux.id
-                self.views.append(aux)
-
-            # Create workspaces, add them the views and store them in self.workspaces
-            for i in range(0, len(self.workspaces_data)):
-                aux = Workspace.objects.create(**self.workspaces_data[i])
-                self.workspaces_data[i]['id'] = aux.id
-                self.workspaces.append(aux)
-                self.workspaces[i].views.add(self.views[i], through_defaults={'view_name': 'v{}'.format(i)})
-                self.workspaces[i].views.add(self.views[i + 1], through_defaults={'view_name': 'v{}'.format(i)})
-        self.old_count = View.objects.count()
-
-    def client_login(self):
-        """Perform a login for the APIClient."""
         data = {'username': self.username, 'password': self.password}
         self.client.post(self.login_url, data, format='json')
         self.token = Token.objects.get(user__username=self.username)
         self.client.credentials(HTTP_AUTHORIZATION='Token ' + self.token.key)
 
-    def test_list_views(self):
-        """Test that the list of views can be retrieved through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='view_view'))
+    def test_authorized_list_objects(self):
+        """Test that authorized users can retrieve the list of objects through the API."""
+        for case in self.cases:
+            # Arrange
+            self.user.user_permissions.add(Permission.objects.get(codename='view_{}'.format(case['key'])))
+            # Act
+            url = reverse('{}-list'.format(case['key']))
+            response = self.client.get(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_200_OK,
+                'Retrieving list of {} did not return status 200'.format(case['class'])
+            )
+            retrieved_data = [dict(data) for data in response.data]
+            self.assertEqual(
+                retrieved_data, case['current_data'],
+                'Retrieved list of {} is not as expected'.format(case['class'])
+            )
 
-        # Act
-        response = self.client.get(reverse('view-list'))
+    def test_authorized_create_objects(self):
+        """Test that authorized users can create objects through the API."""
+        for case in self.cases:
+            # Arrange
+            self.user.user_permissions.add(Permission.objects.get(codename='add_{}'.format(case['key'])))
+            # Act
+            url = reverse('{}-list'.format(case['key']))
+            response = self.client.post(url, case['new_data'])
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_201_CREATED,
+                'Posting a new {} did not return status 201'.format(case['class'])
+            )
+            self.assertEqual(
+                case['class'].objects.count(), case['old_count'] + 1,
+                'The number of {} should have increased by 1'.format(case['class'])
+            )
 
-        # Assert
-        expected_data = [
-            {
-                'id': view['id'],
-                'creation_timestamp': self.setup_ts_str,
-                'update_timestamp': self.setup_ts_str,
-                'name': view['name'],
-                'data': view['data'],
-            } for view in self.views_data
-        ]
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = [dict(v) for v in response.data]
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
+    def test_authorized_retrieve_objects(self):
+        """Test that authorized users can retrieve objects through the API."""
+        for case in self.cases:
+            # Arrange
+            self.user.user_permissions.add(Permission.objects.get(codename='view_{}'.format(case['key'])))
+            # Act
+            obj = case['class'].objects.get(id=case['selected_id'])
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.get(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_200_OK,
+                'Getting a {} did not return status 200'.format(case['class'])
+            )
+            retrieved_data = dict(response.data)
+            self.assertEqual(
+                retrieved_data, case['current_data'][0],
+                'Retrieved list of {} is not as expected'.format(case['class'])
+            )
 
-    def test_create_views(self):
-        """Test that a view can be created through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='add_view'))
-        given_data = {
-            "name": "My New View",
-            "views": [],
-            "data": '{"data_name": "My New View"}'
-        }
+    def test_authorized_update_objects(self):
+        """Test that authorized users can update objects through the API."""
+        for case in self.cases:
+            # Arrange
+            self.user.user_permissions.add(Permission.objects.get(codename='change_{}'.format(case['key'])))
+            # Act
+            obj = case['class'].objects.get(id=case['selected_id'])
+            old_data = get_dict(obj)
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.put(url, case['new_data'])
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_200_OK,
+                'Updating a {} did not return status 200'.format(case['class'])
+            )
+            new_data = get_dict(case['class'].objects.get(pk=obj.pk))
+            self.assertNotEqual(
+                new_data, old_data,
+                'The object {} should have been updated'.format(case['class'])
+            )
 
-        # Act
-        response = self.client.post(reverse('view-list'), given_data)
-
-        # Assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED, 'The request failed')
-        self.new_count = View.objects.count()
-        self.assertEqual(self.old_count + 1, self.new_count, 'A new object should have been created in the DB')
-        new_view = View.objects.get(name=given_data['name'])
-        self.assertEqual(new_view.data, json.loads(given_data['data']), 'Retrieved data is not as expected')
-
-    def test_retrieve_views(self):
-        """Test that a view can be retrieved through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='view_view'))
-        data = self.views_data[0]
-        # Act
-        response = self.client.get(reverse('view-detail', kwargs={'pk': data['id']}))
-
-        # Assert
-        expected_data = {
-            'id': data['id'],
-            'name': data['name'],
-            'creation_timestamp': self.setup_ts_str,
-            'update_timestamp': self.setup_ts_str,
-            'data': data['data'],
-        }
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = dict(response.data)
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
-
-    def test_update_views(self):
-        """Test that a view can be updated through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='change_view'))
-        data = self.views_data[0]
-        given_data = {
-            "name": "My New Workspace",
-            "data": '{"data_name": "My New View"}'
-        }
-        # Act
-        self.update_ts = timezone.now()
-        self.update_ts_str = serializers.DateTimeField().to_representation(self.update_ts)
-        with freeze_time(self.update_ts):
-            response = self.client.put(reverse('view-detail', kwargs={'pk': data['id']}), given_data)
-
-        # Assert
-        expected_data = {
-            'id': data['id'],
-            'name': given_data['name'],
-            'creation_timestamp': self.setup_ts_str,
-            'update_timestamp': self.update_ts_str,
-            'data': json.loads(given_data['data']),
-        }
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = dict(response.data)
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
-
-    def test_delete_views(self):
-        """Test that a view can be deleted through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='delete_view'))
-        view_pk = self.views[0].pk
-        # Act
-        response = self.client.delete(reverse('view-detail', kwargs={'pk': view_pk}))
-
-        # Assert
-        self.new_count = View.objects.count()
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT, 'The request failed')
-        self.assertEqual(self.old_count - 1, self.new_count, 'The number of objects in the DB have decreased by 1')
-        with self.assertRaises(Exception):
-            View.objects.get(pk=view_pk)
-
-
-class WorkspaceViewCrudTestCase(TestCase):
-    """Test the workspace_view CRUD API."""
-
-    def setUp(self):
-        """Testcase setup.
-
-        We start with 3 workspaces and 4 views and we add view_i and view_i+1 to workspace_i
-        """
-        # Arrange
-        # Client setup
-        self.client = APIClient()
-        self.login_url = reverse('login')
-        self.username = 'test'
-        self.password = 'password'
-        self.user = User.objects.create_user(
-            username=self.username,
-            password='password',
-            email='test@user.cl',
-            first_name='First',
-            last_name='Last',
-        )
-        self.setup_ts = timezone.now()
-        self.setup_ts_str = serializers.DateTimeField().to_representation(self.setup_ts)
-
-        with freeze_time(self.setup_ts):
-            self.views_data = [
-                {
-                    'name': 'My View 1',
-                    'data': json.dumps({"data_name": "My View 1"}),
-                },
-                {
-                    'name': 'My View 2',
-                    'data': json.dumps({"data_name": "My View 2"}),
-                },
-                {
-                    'name': 'My View 3',
-                    'data': json.dumps({"data_name": "My View 3"}),
-                },
-                {
-                    'name': 'My View 4',
-                    'data': json.dumps({"data_name": "My View 4"}),
-                }
-            ]
-            self.workspaces_data = [
-                {'name': 'My Workspace 1'},
-                {'name': 'My Workspace 2'},
-                {'name': 'My Workspace 3'},
-            ]
-            self.views = []
-            self.workspaces = []
-            self.workspace_views_data = []
-            # Create views and store them in self.views
-            for i in range(0, len(self.views_data)):
-                aux = View.objects.create(**self.views_data[i])
-                self.views_data[i]['id'] = aux.id
-                self.views.append(aux)
-
-            # Create workspaces, add them the views and store them in self.workspaces,
-            # and store data from workspace_views in self.workspace_views_data for asserting
-            for i in range(0, len(self.workspaces_data)):
-                aux = Workspace.objects.create(**self.workspaces_data[i])
-                self.workspaces_data[i]['id'] = aux.id
-                self.workspaces.append(aux)
-                # Add view[i]
-                self.workspaces[i].views.add(self.views[i], through_defaults={'view_name': 'v{}'.format(i)})
-                aux = WorkspaceView.objects.get(workspace=self.workspaces[i], view=self.views[i])
-                self.workspace_views_data.append({
-                    'id': aux.id,
-                    'creation_timestamp': self.setup_ts_str,
-                    'update_timestamp': self.setup_ts_str,
-                    'view_name': 'v{}'.format(i),
-                    'sort_value': 0,
-                    'view': self.views[i].pk,
-                    'workspace': self.workspaces[i].pk,
-                })
-                # Add view[i + 1]
-                self.workspaces[i].views.add(self.views[i + 1], through_defaults={'view_name': 'v{}'.format(i)})
-                aux = WorkspaceView.objects.get(workspace=self.workspaces[i], view=self.views[i + 1])
-                self.workspace_views_data.append({
-                    'id': aux.id,
-                    'creation_timestamp': self.setup_ts_str,
-                    'update_timestamp': self.setup_ts_str,
-                    'view_name': 'v{}'.format(i),
-                    'sort_value': 0,
-                    'view': self.views[i + 1].pk,
-                    'workspace': self.workspaces[i].pk,
-                })
-        self.old_count = WorkspaceView.objects.count()
-
-    def client_login(self):
-        """Perform a login for the APIClient."""
-        data = {'username': self.username, 'password': self.password}
-        self.client.post(self.login_url, data, format='json')
-        self.token = Token.objects.get(user__username=self.username)
-        self.client.credentials(HTTP_AUTHORIZATION='Token ' + self.token.key)
-
-    def test_list_workspace_views(self):
-        """Test that the list of workspace_views can be retrieved through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='view_workspaceview'))
-
-        # Act
-        response = self.client.get(reverse('workspaceview-list'))
-
-        # Assert
-        expected_data = self.workspace_views_data
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = [dict(v) for v in response.data]
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
-
-    def test_create_worskpace_views(self):
-        """Test that a workspace_view can be created through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='add_workspaceview'))
-        given_data = {
-            "view_name": "My New Name",
-            "workspace": self.workspaces[0].id,
-            "view": self.views[3].id,
-            "sort_value": 1,
-        }
-
-        # Act
-        response = self.client.post(reverse('workspaceview-list'), given_data)
-
-        # Assert
-        self.assertEqual(response.status_code, status.HTTP_201_CREATED, 'The request failed')
-        self.new_count = WorkspaceView.objects.count()
-        self.assertEqual(self.old_count + 1, self.new_count, 'A new object should have been created in the DB')
-        new_wv = WorkspaceView.objects.get(view_name=given_data['view_name'])
-        self.assertEqual(new_wv.view_name, given_data['view_name'], 'New objects view_name is not as expected')
-        self.assertEqual(new_wv.view.id, given_data['view'], 'New objects view is not as expected')
-        self.assertEqual(new_wv.workspace.id, given_data['workspace'], 'New objects workspace is not as expected')
-        self.assertEqual(new_wv.sort_value, given_data['sort_value'], 'New objects sort_value is not as expected')
-
-    def test_retrieve_workspace_views(self):
-        """Test that a workspace_views can be retrieved through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='view_workspaceview'))
-        data = self.workspace_views_data[0]
-        # Act
-        response = self.client.get(reverse('workspaceview-detail', kwargs={'pk': data['id']}))
-
-        # Assert
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = dict(response.data)
-        self.assertEqual(retrieved_data, data, 'Retrieved data is not as expected')
-
-    def test_update_workspace_view(self):
-        """Test that a workspace_view can be updated through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='change_workspaceview'))
-        data = self.workspace_views_data[0]
-        given_data = {
-            "view_name": "My New Name",
-            'view': self.views[3].id,
-            'workspace': self.workspaces[1].id,
-        }
-        # Act
-        self.update_ts = timezone.now()
-        self.update_ts_str = serializers.DateTimeField().to_representation(self.update_ts)
-        with freeze_time(self.update_ts):
-            response = self.client.put(reverse('workspaceview-detail', kwargs={'pk': data['id']}), given_data)
-
-        # Assert
-        expected_data = {
-            'id': data['id'],
-            'creation_timestamp': self.setup_ts_str,
-            'update_timestamp': self.update_ts_str,
-            'view_name': given_data['view_name'],
-            'sort_value': 0,
-            'view': self.views[3].id,
-            'workspace': self.workspaces[1].id,
-        }
-        self.assertEqual(response.status_code, status.HTTP_200_OK, 'The request failed')
-        retrieved_data = dict(response.data)
-        self.assertEqual(retrieved_data, expected_data, 'Retrieved data is not as expected')
-
-    def test_delete_workspace_views(self):
-        """Test that a workspace_view can be deleted through the API."""
-        # Arrange
-        self.client_login()
-        self.user.user_permissions.add(Permission.objects.get(codename='delete_workspaceview'))
-        wv_pk = self.workspace_views_data[0]['id']
-        # Act
-        response = self.client.delete(reverse('workspaceview-detail', kwargs={'pk': wv_pk}))
-
-        # Assert
-        self.new_count = WorkspaceView.objects.count()
-        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT, 'The request failed')
-        self.assertEqual(self.old_count - 1, self.new_count, 'The number of objects in the DB have decreased by 1')
-        with self.assertRaises(Exception):
-            WorkspaceView.objects.get(pk=wv_pk)
+    def test_authorized_delete_objects(self):
+        """Test that authorized users can dalete objects through the API."""
+        for case in self.cases:
+            # Arrange
+            old_count = case['class'].objects.count()
+            self.user.user_permissions.add(Permission.objects.get(codename='delete_{}'.format(case['key'])))
+            # Act
+            obj = case['class'].objects.first()
+            url = reverse('{}-detail'.format(case['key']), kwargs={'pk': obj.pk})
+            response = self.client.delete(url)
+            # Assert
+            self.assertEqual(
+                response.status_code, status.HTTP_204_NO_CONTENT,
+                'Deleting a {} did not return status 204'.format(case['class'])
+            )
+            self.assertEqual(
+                case['class'].objects.count(), old_count - 1,
+                'The number of {} should have decreased by 1'.format(case['class'])
+            )
